Start to move built in types into tests
