+++ smol-core/test/Test/IR/DataTypesSpec.hs
@@ -8,7 +8,7 @@ import qualified Data.Map.Strict as M
 import qualified Smol.Core.IR.FromExpr.DataTypes as DT
 import Smol.Core.IR.FromExpr.Types
 import Smol.Core.IR.FromResolvedExpr
-import qualified Smol.Core.Typecheck as TC
+import Smol.Core.Typecheck.BuiltInTypes (builtInTypes)
 import Smol.Core.Typecheck.FromParsedExpr
 import qualified Smol.Core.Typecheck.Types as Smol
 import Smol.Core.Types
@@ -22,7 +22,7 @@ typeToDataType ::
 typeToDataType ty =
   evalState
     (DT.typeToDataTypeInMemory ty)
-    (FromExprState mempty (TC.builtInTypes Identity) 0 mempty)
+    (FromExprState mempty (builtInTypes Identity) 0 mempty)
 
 parseToIdentity :: Type ParseDep a -> Type Identity a
 parseToIdentity = fromResolvedType . fromParsedType
