+++ smol-core/test/Test/IR/FromExprSpec.hs
@@ -15,13 +15,14 @@ import Smol.Core.IR.FromResolvedExpr
 import Smol.Core.IR.IRExpr
 import Smol.Core.IR.ToLLVM.Patterns
 import Smol.Core.Typecheck
+import Smol.Core.Typecheck.BuiltInTypes (builtInTypes)
 import Smol.Core.Types
 import Test.Helpers
 import Test.Hspec
 
 evalExpr :: Text -> IdentityExpr (Type Identity Annotation)
 evalExpr input =
-  case elaborate (unsafeParseTypedExpr input $> mempty) of
+  case elaborate (builtInTypes emptyResolvedDep) (unsafeParseTypedExpr input $> mempty) of
     Right typedExpr -> fromResolvedType <$> fromResolvedExpr typedExpr
     Left e -> error (show e)
