+++ smol-core/test/Test/Typecheck/ExhaustivenessSpec.hs
@@ -9,6 +9,7 @@ import Control.Monad.Except
 import Control.Monad.Reader
 import qualified Data.List.NonEmpty as NE
 import Smol.Core
+import Smol.Core.Typecheck.BuiltInTypes
 import Smol.Core.Typecheck.FromParsedExpr
 import Test.Helpers
 import Test.Hspec
